refactor(capi): extract appearance model to CAPI_appearance.cpp (todo 2.4)
First multi-TU split step. Moved verbatim: companion lookup, effective
appearance, resolve/persist logic, and the 11 exported appearance
functions (RFC 0007). The 3 helpers shared with CAPI.cpp (dasher_create,
dasher_save_settings) are declared in CAPI_internal.h (namespace capi);
everything else is TU-local. pugixml include dropped from CAPI.cpp.

All three compile sites updated: dasher shared-lib target,
dasher_control_action_tests (compiles CAPI.cpp directly), release.yml
WASM em++ link. CAPI.cpp: 2059 -> 1839 lines.
Verified: build [100%] zero errors (full log), 45/45 ctest, freeze 110.
