+++ .github/workflows/release.yml
@@ -169,6 +169,7 @@ jobs:
           WEB=dasher-web/wasm-build
           em++ -O3 -DNDEBUG -fwasm-exceptions \
               src/CAPI.cpp \
+              src/CAPI_appearance.cpp \
               -I src \
               -I Thirdparty/pugixml/src \
               build-wasm/libDasherCore.a \
